requirements.txt
huggingface-hub==0.24.6
numpy==2.0.1
pandas==2.2.2
requests==2.32.2
yfinance==0.2.41
fredapi==0.5.2
schedule==1.2.2
tqdm==4.66.4
uvicorn
fastapi
mplfinance
jinja2
python-dotenv

